requirements.txt
paho-mqtt
coloredlogs
tqdm
python-dotenv